@@ -20,34 +20,34 @@
         }
     },
     "require-dev": {
-        "symfony/debug": "^3.1",
-        "phpunit/phpunit": "^6",
+        "symfony/debug": "^4",
+        "phpunit/phpunit": "^5",
         "phpunit/php-invoker": "^2",
-        "phpunit/php-code-coverage": "^5",
-        "components/bootstrap-default": "^3.3",
-        "symfony/browser-kit": "^3.1",
-        "symfony/event-dispatcher": "^4",
-        "symfony/filesystem": "^4",
-        "symfony/process": "^4",
-        "there4/slim-test-helpers": "^2.1",
-        "symfony/flex": "^1.1"
+        "phpunit/php-code-coverage": "4",
+        "symfony/browser-kit": "^5",
+        "symfony/event-dispatcher": "^5",
+        "symfony/filesystem": "^5",
+        "symfony/process": "^5",
+        "there4/slim-test-helpers": "^2",
+        "symfony/flex": "^1"
     },
     "require": {
-        "php": "^7.0",
-        "slim/slim": "^3.0",
-        "slim/twig-view": "~1.0",
-        "slim/flash": "^0.1.0",
-        "monolog/monolog": "^1.13",
+        "php": "^7.3",
+        "slim/slim": "~3.1",
+        "slim/twig-view": "^2",
+        "slim/flash": "^0.4",
+        "monolog/monolog": "^2",
         "codeguy/upload": "^1.3",
-        "illuminate/database": "~5.1",
-        "ircmaxell/password-compat": "^1.0",
-        "phpoffice/phpspreadsheet": "^1.4",
-        "symfony/console": "^4.1",
+        "illuminate/database": "^6",
+        "ircmaxell/password-compat": "^1",
+        "phpoffice/phpspreadsheet": "^1",
+        "symfony/console": "^5",
         "ext-json": "^1.6",
         "ext-PDO": "^7.2",
-        "symfony/dom-crawler": "^4.2",
+        "symfony/dom-crawler": "^5",
         "ext-curl": "^7.3",
-        "symfony/css-selector": "^4.2",
-        "joshfraser/php-name-parser": "dev-master"
+        "symfony/css-selector": "^5",
+        "joshfraser/php-name-parser": "dev-master",
+        "components/bootstrap-default": "^3.3"
     }
 }
